Clean-up loading of dynamic libraries
