+++ src/pyft4222/wrapper/i2c/slave.py
@@ -5,7 +5,7 @@ from typing import NewType
 
 from pyft4222.result import Err, Ok, Result
 from pyft4222.wrapper import Ft4222Exception, Ft4222Status, FtHandle
-from pyft4222.wrapper.dll_loader import ftlib
+from pyft4222.wrapper.dll import ftlib
 
 I2cSlaveHandle = NewType("I2cSlaveHandle", FtHandle)
